requirements.txt
# ============================================================================
# Delta-Neutral Hedge Trading Bot - Dependencies
# Self-Contained Version
# ============================================================================

# ============= Exchange APIs =============
# Bybit official Python SDK
pybit==5.6.2

# HTTP requests for API calls
requests>=2.28.0

# WebSocket support
websocket-client>=1.3.0

# ============= Database =============
# PostgreSQL adapter
psycopg2-binary>=2.9.0

# ============= Redis (Price Feed) =============
# Redis client for price data
redis>=5.0.1

# ============= Configuration =============
# Environment variable management
python-dotenv>=1.0.0

# ============= Async Support =============
# Required for WebSocket connections
aiohttp>=3.8.0
python-socketio[asyncio]>=5.7.0

# ============= Data Handling =============
# JSON handling (stdlib, but listed for clarity)
# json (built-in)

# ============= Optional: Development & Testing =============
# Uncomment if you want to run tests
# pytest>=7.4.0
# pytest-asyncio>=0.21.0

# ============= Notes =============
# This requirements.txt includes ALL dependencies needed for the bot
# No external code linking required - completely self-contained
#
# Installation:
#   pip install -r requirements.txt
#
# Python Version Required: 3.8+
